// Neptune.Web/src/app/shared/generated/model/water-quality-management-plan-verify-upsert-dto.ts
export class WaterQualityManagementPlanVerifyUpsertDto { 
    WaterQualityManagementPlanVerifyTypeID: number;
    WaterQualityManagementPlanVisitStatusID: number;
    VerificationDate: string;
    WaterQualityManagementPlanVerifyStatusID?: number | null;
    SourceControlCondition?: string | null;
    EnforcementOrFollowupActions?: string | null;
    IsDraft?: boolean;
    TreatmentBMPs?: Array<VerifyTreatmentBMPUpsertDto> | null;
    QuickBMPs?: Array<VerifyQuickBMPUpsertDto> | null;
    SourceControlBMPs?: Array<VerifySourceControlBMPUpsertDto> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}